refactor: extract delivery pr metadata
